pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bwenv"
version = "1.0.0"
authors = [
  { name="Arkin Modi", email="arkinmodi@gmail.com" },
]
description = "Securely use and share environment variables in your local development using Bitwarden"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=0.8.0"
]

[project.scripts]
bwenv = "bw_env._main:main"

[project.urls]
"Homepage" = "https://github.com/arkinmodi/bitwarden-env"

[tool.setuptools.packages.find]
exclude = ["assets*", "tests*"]
